--- front/src/providers/ChakraProviders.tsx
@@ -1,12 +1,0 @@
-import { CacheProvider } from '@chakra-ui/next-js'
-import { ChakraProvider } from '@chakra-ui/react'
-
-export function Providers({ children }: { children: React.ReactNode }) {
-    return (
-        <CacheProvider>
-            <ChakraProvider>
-                {children}
-            </ChakraProvider>
-        </CacheProvider>
-    )
-}
